requirements.txt
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.20
crewai>=0.11.0
pydantic>=2.5.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
streamlit>=1.28.1
pandas>=2.1.3
numpy>=1.26.0
matplotlib>=3.8.2
seaborn>=0.13.0
requests>=2.31.0
aiohttp>=3.9.1
setuptools>=68.0.0
wheel>=0.40.0
plotly>=5.0.0 